+++ src/modules/symbols/numero-sign.js
@@ -1,9 +1,8 @@
 import { base } from "../../const.js";
 import { fixSpacingAroundSymbol } from "./symbol-utils.js";
 
-
-export function fixNumeroSign(string) {
-  string = fixSpacingAroundSymbol(string, base.numeroSign);
+export function fixNumeroSign(string, locale) {
+  string = fixSpacingAroundSymbol(string, base.numeroSign, locale.spaceAfter.numeroSign);
 
   return string;
 }